Remove "widgets" from Windows
Disable a setting called "Allow widgets".  This shows news and things on the taskbar...but presumably only when the device can reach the service that provides the data.  These settings manifest as soccer balls, news, stock tickers, or whatnot on the taskbar and in the Search section on the taskbar.  I believe this started with 22H2 in Windows 10.  Also applies to Windows 11.