Update "remove OneDrive" code
Addressing an issue where the current VDOT "remove OneDrive" code was run, and it worked as intended.  Yet OneDrive still shows up on the Start menu.  It was not functional of course, but where customers ask for OneDrive to be removed, they don't want it right there on the Start menu.  This code does a more sweeping deletion of OneDrive files.
And it's not like you can't get it back if you want it.  You can download the OneDrive installer.